Start Student React component
Students will need a webpage(s) to display courses they can join and
have already joined.  Further discussion is needed on the styling and
layout of these pages, so this commit is primarily intended as a proof
of concept.

The UI's styling is rendered by react-bootstrap.  react-bootstrap is
a complete re-implementation of Bootstrap components in React.  Like
Bootstrap, it defines many CSS classes to make pretty UI elements and
uses flexbox to make flexible layouts.  More info can be found at
https://bre.is/JX2hfmU4.

Issues: #34