Rename uuid and remove disabled Join buttons
The agreement reached through group discussions is that UUIDs are not
needed at this time.  A student's ID is used to distinguish the student
from other students, and not from other entity types, such as courses.
Other entity types will have their own pools of IDs to assign IDs from,
so student IDs do not need to be universally unique across entity
types.  Therefore, it is sufficient to rename the "uuid" field to "id".

The Student webpage is now implemented under the assumption that it
represents a student's profile page.  This means the displayed course
list is the list of courses the student is enrolled in.  Furthermore,
in this context, the term "Join" is defined as entering a class
session, not as enrolling in a course.  Therefore, disabled Join
buttons are no longer appropriate.