Start Course Session React Component
Instructors will need a webpage(s) to display the questions associated
with a course session. Further discussion is needed on the styling and
layout of these pages, so this commit is primarily intended as a proof
of concept.

The UI's styling is rendered by react-bootstrap. react-bootsrap is a
complete re-implementation of Boostrap components in React. Like
Bootstrap, it defines many CSS classes to make pretty UI elements and
uses flexbox to make flexibly layouts. More info can be found at
https://bre.is/JX2hfmU4.

Issues: #32